% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmParallel.R
\name{dtm_original}
\alias{dtm_original}
\alias{dtm_parallel_v1}
\alias{dtm_parallel_v2}
\title{hunspell Spell Checking and Morphological Analysis}
\usage{
dtm_original(texts, min_term_freq = 0L,
  max_term_freq = .Machine$integer.max, min_word_length = 0L,
  max_word_length = .Machine$integer.max,
  threadsCount = RcppParallel::defaultNumThreads())

dtm_parallel_v1(texts, min_term_freq = 0L,
  max_term_freq = .Machine$integer.max, min_word_length = 0L,
  max_word_length = .Machine$integer.max,
  threadsCount = RcppParallel::defaultNumThreads())

dtm_parallel_v2(texts, min_term_freq = 0L,
  max_term_freq = .Machine$integer.max, min_word_length = 0L,
  max_word_length = .Machine$integer.max,
  threadsCount = RcppParallel::defaultNumThreads())
}
\description{
hunspell Spell Checking and Morphological Analysis
}

