% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/performanceTests.R
\name{DocumentTermMatrixOriginal}
\alias{DocumentTermMatrixOriginal}
\alias{DocumentTermMatrixParallelSlower}
\alias{TermDocumentMatrixOriginal}
\alias{TermDocumentMatrixParallelSlower}
\title{tmParallel functions for performance testing}
\usage{
DocumentTermMatrixOriginal(texts, control = list())

TermDocumentMatrixOriginal(texts, control = list())

DocumentTermMatrixParallelSlower(texts, control = list())

TermDocumentMatrixParallelSlower(texts, control = list())
}
\description{
Some functions usefull for performance testing
}

