% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weightParallel.R
\name{weightTfIdfParallel}
\alias{weightTfIdfParallel}
\title{weightTfIdfParallel}
\usage{
weightTfIdfParallel(m, normalize = TRUE, plus1inIdf = FALSE)
}
\arguments{
\item{m}{either \code{\link{TextDocumentMatrix}}
or a \code{\link{DocumentTextMatrix}} object that extens
\code{\link{slam}} \code{\link[slam]{simple_triplet_matrix}}.}

\item{normalize}{logical flag, if \code{TRUE} then Tf-Idf is normalized
- words count vector is divided by words count in document in which this word is from.
Default \code{TRUE}.}

\item{plus1inIdf}{logical flag, if \code{TRUE} then when counting Idf as
\code{log2( documents count / documents with given word )} the costant \code{1} value
is added.
Default \code{FALSE}.}
}
\value{
The given \code{\link{DocumentTermMatrix}} or \code{\link{TermDocumentMatrix}}
with updated terms frequency values.
}
\description{
Parallel iplementatnion of Tf-Idf weighting function
}
\examples{
pdtm <-
 ParallelDocumentTermMatrix(c('This is a test', 'And another test') )
w1pdtm <-
 weightTfIdfParallel(pdtm, normalize = TRUE, plus1inIdf = FALSE)
w2pdtm <-
 weightTfIdfParallel(pdtm, normalize = FALSE, plus1inIdf = TRUE)

pdtm <-
 ParallelDocumentTermMatrix(
  c('This is a test test', 'And another test'),
  control = list( weighting = weightTfIdfParallel )
 )

pdtm <-
 ParallelDocumentTermMatrix(
  c('This is a test test', 'And another test'),
  control = list(
   weighting = function(x) weightTfIdfParallel(x, normalize = TRUE, plus1inIdf = FALSE)
  )
 )
}
\seealso{
\code{\link{tm}} \code{\link[tm]{weighting}}
 for other weighting functions.
}

